Chức năng tạo báo cáo chi phí

+++ DTO/Tbl_DTO/tbl_Report_Expense_DTO.cs
@@ -11,7 +11,7 @@ namespace DTO.tbl_DTO
     /// </summary>
     public class tbl_Report_Expense_DTO
     {
-        public decimal TotalExpenses { get; set; }
+        public double TotalExpenses { get; set; }
         public int NumberOfExpenses { get; set; }
     }
 }
